Patch 'File IO' examples so they don't break
Examples trying to save files break because they are located in /app
which is read-only.  Ideally the examples directory should be moved to a
writtable location, but Processing expects them where they are now.  So
the example pde files are patched instead by changing the path from
relative 'foo' to absolute '/tmp/foo'.